requirements.txt
Flask>=2.0
psycopg2-binary>=2.9
gunicorn>=20.1
python-dotenv>=1.0